deduplicate dispatchers
